Rename module dataset_populator to dataset_item_uploader
